backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background Tasks
celery==5.3.4
redis==5.0.1

# File Processing
PyPDF2==3.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==11.2.1

# AI/LLM Integration
openai==1.40.0
httpx==0.25.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0

#LangChain Framework
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.16
langchain-core==0.2.40
